+++ src/community/people/api/utils/ApiEndpoints.ts
@@ -22,7 +22,9 @@ export const peoplesEndpoints = {
     `${moduleAPIPath.PEOPLE}/user/terminate/${employeeId}`,
   SEARCH_EMPLOYEE_TEAM_ADMIN: `${moduleAPIPath.PEOPLE}/search/employee-team`,
   CHECK_IF_CURRENT_USER_HAS_MANAGERS_AVAILABILITY: `${moduleAPIPath.PEOPLE}/me/managers/availability`,
-  MY_MANAGERS: `${moduleAPIPath.PEOPLE}/me/managers`
+  MY_MANAGERS: `${moduleAPIPath.PEOPLE}/me/managers`,
+  SUPERVISED_BY_ME: (employeeId: number) =>
+    `${moduleAPIPath.PEOPLE}/${employeeId}/is-supervised-by-me`
 };
 
 export const authEndpoints = {
