+++ src/community/people/api/utils/QueryKeys.ts
@@ -53,7 +53,8 @@ export const peopleQueryKeys = {
       workEmail,
       identificationNo
     ].filter((val) => val !== undefined);
-  }
+  },
+  SUPERVISED_BY_ME: ["supervised-by-me"]
 };
 
 export const teamQueryKeys = {
@@ -68,8 +69,7 @@ export const teamQueryKeys = {
     return [...(this?.all || []), "getSearchedTeams", searchedTeams].filter(
       (val) => val !== undefined
     );
-  },
-  SUPERVISED_BY_ME: ["supervised-by-me"]
+  }
 };
 
 export const jobFamilyQueryKeys = {
